Added code that corrects a problem with loading the data.
here is the comment:
This code fixes https://github.com/aws-amplify/amplify-js/issues/13267 .
Which is considered a bug.  When a new record is inserted into GraphQL,
the item returned to the observable does not have any of its nested items
populated.  This code finds items that have item.status as null.  It
looks in the GraphQL database and replaces those items with good ones.